you need to set terror to isFollowing to make it sit
MC hardcoded the riding and the vehicle returns a normal entity instead of a synced compoundTag. this means if the player disconnects the terror would dismount. but the effects have a 1 minute timer so the player is saved but the dragon falls. to alleviate this. the terrror needs to be set to following so it floats above player after disconnecting.